Addition of basic instrumentation, and file based load / unload of SMA Note that this "works" with the PHP test suite on the zend dir.  The reason for the quotes is that it doesn't crash, but there are still bugs to be shaken down.
The scope of the "instrumentation" fork is to act as a "proof of principle" of
file based caching.  It basically adds persistence to cli (and later gci) based
opcache for a limited test scenario.  The immediate goals are:
  1) To have a functional memory-based cache albeit with known limitations
  2) To exercise this on the PHP tests suite
  3) As a basline to do further development of the concept.

Basically as the design docs for O+ are the source -- fin -- static analysis
can be misleading.  To do dynamic analysis you need the type of stepping stone
verson :-(

I've committed this snapshot as a WIP to permit third party comment.  Instead of
maintaining TODO and buttlist docs within the repository, I'll add these to the
public Git wiki.